+++ 1.12/src/main/java/panszelescik/coolgadgets/CoolGadgets.java
@@ -12,11 +12,12 @@ import org.apache.logging.log4j.Logger;
 import panszelescik.coolgadgets.helper.BotaniaClientHelper;
 import panszelescik.coolgadgets.helper.BotaniaHelper;
 import panszelescik.coolgadgets.helper.CraftingHelper;
+import panszelescik.coolgadgets.helper.EnvironmentalTechHelper;
 import panszelescik.coolgadgets.items.ModItems;
 import panszelescik.coolgadgets.proxy.CommonProxy;
 import panszelescik.morelibs.api.Helper;
 
-@Mod(modid = CoolGadgets.MODID, name = "Cool Gadgets", version = "@version@", acceptedMinecraftVersions = "[1.12.2,1.13)", updateJSON = "https://raw.githubusercontent.com/PanSzelescik/coolgadgets/master/update.json", dependencies = "required-after:morelibs@[1.2.0,);after:patchouli@[1.0-13,);after:actuallyadditions;after:appliedenergistics2;after:botania;after:buildcraftcore;after:cofhcore;after:draconicevolution;after:enderio;after:extrautils2;after:factorytech;after:hammercore;after:ic2;after:immersiveengineering;after:integrateddynamics;after:mcjtylib;after:mekanism;after:opencomputers;after:pneumaticcraft;after:projectred-core;after:reborncore;after:refinedstorage")
+@Mod(modid = CoolGadgets.MODID, name = "Cool Gadgets", version = "@version@", acceptedMinecraftVersions = "[1.12.2,1.13)", updateJSON = "https://raw.githubusercontent.com/PanSzelescik/coolgadgets/master/update.json", dependencies = "required-after:morelibs@[1.2.0,);after:patchouli@[1.0-13,);after:actuallyadditions;after:appliedenergistics2;after:botania;after:buildcraftcore;after:cofhcore;after:draconicevolution;after:enderio;after:extrautils2;after:factorytech;after:hammercore;after:ic2;after:immersiveengineering;after:integrateddynamics;after:jei;after:mcjtylib;after:mekanism;after:opencomputers;after:pneumaticcraft;after:projectred-core;after:reborncore;after:refinedstorage;after:valkyrielib")
 public class CoolGadgets {
 
     public static final String MODID = "coolgadgets";
@@ -48,6 +49,9 @@ public class CoolGadgets {
         if (Helper.isLoaded("botania")) {
             MinecraftForge.EVENT_BUS.register(BotaniaHelper.class);
             MinecraftForge.EVENT_BUS.register(BotaniaClientHelper.class);
+        }
+        if (Helper.isLoaded("valkyrielib")) {
+            MinecraftForge.EVENT_BUS.register(EnvironmentalTechHelper.class);
         }
     }
 }